Split policy-ladder admission and finish out of the tier runner.
Keep the step module as orchestration only so early-stop and diagnostics stop growing inside the solve loop.
