+++ packages/api/api/analytics/military_score_inference/policy_ladder_tier_step.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 
 import time
 from collections.abc import Callable, Sequence
-from enum import Enum
 from pathlib import Path
 
 from api.analytics.military_score_inference.actions import (
@@ -19,10 +18,6 @@ from api.analytics.military_score_inference.collision_hull_widen import (
 )
 from api.analytics.military_score_inference.component_eligibility import (
     player_by_id,
-    turn_catalog_context_for_policy_step,
-)
-from api.analytics.military_score_inference.constraints import (
-    solution_satisfies_exact_hard_equalities,
 )
 from api.analytics.military_score_inference.degrade_aggregate_probe import (
     probe_degrade_aggregate_rewrites,
@@ -34,6 +29,9 @@ from api.analytics.military_score_inference.models import (
     InferenceResult,
     InferenceSolution,
 )
+from api.analytics.military_score_inference.policy_ladder_admission import (
+    _make_incremental_admitter,
+)
 from api.analytics.military_score_inference.policy_ladder_state import PolicyLadderState
 from api.analytics.military_score_inference.policy_ladder_tier_budget import (
     TierStepRun,
@@ -41,58 +39,24 @@ from api.analytics.military_score_inference.policy_ladder_tier_budget import (
     remaining_time,
     tier_step_allowance_seconds,
 )
+from api.analytics.military_score_inference.policy_ladder_tier_finish import (
+    TierStepFinishMode,
+    finish_tier_step,
+)
 from api.analytics.military_score_inference.prior_fleet_tech_raise import (
-    PriorFleetTechRaisePlan,
     resolve_prior_fleet_tech_raise_plan,
 )
 from api.analytics.military_score_inference.solver import (
     STATUS_INVALID_PROBLEM,
     STATUS_STOPPED,
     STATUS_TIME_LIMITED,
-    solution_signature,
     solve_inference_problem,
-)
-from api.analytics.military_score_inference.tier_emission_ledger import (
-    tier_emission_fields,
-)
-from api.analytics.military_score_inference.tier_policy import (
-    InferenceTierPolicyStep,
-    resolve_solver_thresholds,
 )
 from api.models.game import TurnInfo
 
 
 def _combo_counts_from_solution(solution: InferenceSolution) -> dict[str, int]:
     return {ship_build.combo_id: ship_build.count for ship_build in solution.ship_builds}
-
-
-def _best_merged_solution(
-    merged_solutions: list[InferenceSolution],
-) -> InferenceSolution | None:
-    if not merged_solutions:
-        return None
-    return max(merged_solutions, key=lambda solution: solution.objective_value)
-
-
-def _solution_fully_explained_by_ship_builds_only(
-    solution: InferenceSolution,
-    observation: InferenceObservation,
-    catalog: ActionCatalog,
-) -> bool:
-    if solution.actions:
-        return False
-    return solution_satisfies_exact_hard_equalities(solution, observation, catalog)
-
-
-def _solution_qualifies_for_ship_only_exact_early_stop(
-    solution: InferenceSolution,
-    observation: InferenceObservation,
-    catalog: ActionCatalog,
-) -> bool:
-    if not _solution_fully_explained_by_ship_builds_only(solution, observation, catalog):
-        return False
-    thresholds = resolve_solver_thresholds()
-    return solution.objective_value >= thresholds.ship_only_exact_early_stop_min_plausibility
 
 
 def _explained_military_score_2x(
@@ -109,42 +73,6 @@ def _explained_military_score_2x(
         combo = combos_by_id[ship_build.combo_id]
         explained += combo.score_delta_2x * ship_build.count
     return explained
-
-
-def _merge_exact_solutions(
-    merged_solutions: list[InferenceSolution],
-    seen_signatures: set[tuple[tuple[str, int], ...]],
-    candidates: tuple[InferenceSolution, ...],
-    *,
-    resolved_max_solutions: int,
-    on_admitted: Callable[[InferenceSolution], None] | None = None,
-) -> int:
-    new_solutions = 0
-    for solution in candidates:
-        signature = solution_signature(solution)
-        if signature in seen_signatures:
-            continue
-        if len(merged_solutions) < resolved_max_solutions:
-            seen_signatures.add(signature)
-            merged_solutions.append(solution)
-            new_solutions += 1
-            if on_admitted is not None:
-                on_admitted(solution)
-            continue
-        worst_index = min(
-            range(len(merged_solutions)),
-            key=lambda index: merged_solutions[index].objective_value,
-        )
-        worst_solution = merged_solutions[worst_index]
-        if solution.objective_value <= worst_solution.objective_value:
-            continue
-        seen_signatures.remove(solution_signature(worst_solution))
-        seen_signatures.add(signature)
-        merged_solutions[worst_index] = solution
-        new_solutions += 1
-        if on_admitted is not None:
-            on_admitted(solution)
-    return new_solutions
 
 
 def _solve_catalog(
@@ -257,57 +185,6 @@ def _solve_seed_progression(
     return None, None
 
 
-def _policy_step_diagnostics(
-    *,
-    policy_step: InferenceTierPolicyStep,
-    policy_step_index: int,
-    catalog: ActionCatalog | None,
-    turn: TurnInfo,
-    observation: InferenceObservation,
-    seed_count: int,
-    band_residual_2x: int | None,
-    emission_fields: dict[str, object] | None = None,
-    collision_widen: CollisionHullWidenPlan | None = None,
-    prior_fleet_tech_raise: PriorFleetTechRaisePlan | None = None,
-) -> dict[str, object]:
-    catalog_context = turn_catalog_context_for_policy_step(
-        turn,
-        observation.player_id,
-        policy_step,
-    )
-    diagnostics: dict[str, object] = {
-        "policyStepId": policy_step.id,
-        "policyStepIndex": policy_step_index,
-        "policyStepsAttempted": policy_step_index + 1,
-        "constraintSnapshot": policy_step.constraint_snapshot(),
-        "resolvedEligibleEngineIds": sorted(catalog_context.eligible_engine_ids),
-        "resolvedEligibleBeamIds": sorted(catalog_context.eligible_beam_ids),
-        "resolvedEligibleTorpIds": sorted(catalog_context.eligible_torp_ids),
-        "resolvedBuildableHullIds": sorted(catalog_context.buildable_hull_ids),
-        "alpha": policy_step.alpha,
-        "comboCount": len(catalog.ship_build_combos) if catalog is not None else 0,
-        "seedCount": seed_count,
-        "bandResidual2x": band_residual_2x,
-        "allowShipOnlyExactEarlyStop": policy_step.allow_ship_only_exact_early_stop,
-        "hullCollisionTwinWiden": policy_step.hull_collision_twin_widen,
-        "runDegradeAggregateProbe": policy_step.run_degrade_aggregate_probe,
-    }
-    if emission_fields is not None:
-        diagnostics.update(emission_fields)
-    if collision_widen is not None:
-        diagnostics.update(collision_widen.to_diagnostics())
-    if prior_fleet_tech_raise is not None:
-        diagnostics.update(prior_fleet_tech_raise.to_diagnostics())
-    return diagnostics
-
-
-def _annotate_last_step_early_stop(state: PolicyLadderState) -> None:
-    reason = state.ladder_early_stop_reason
-    if reason is None or not state.step_diagnostics:
-        return
-    state.step_diagnostics[-1]["ladderEarlyStopReason"] = reason
-
-
 def _ensure_hull_collision_twins_loaded(state: PolicyLadderState, turn: TurnInfo) -> None:
     if state.hull_collision_twins_loaded:
         return
@@ -316,174 +193,6 @@ def _ensure_hull_collision_twins_loaded(state: PolicyLadderState, turn: TurnInfo
     state.hull_collision_twins_path = str(path) if path is not None else None
     state.hull_collision_twins_fell_back = fell_back
     state.hull_collision_twins_loaded = True
-
-
-def _maybe_early_stop_after_step(
-    state: PolicyLadderState,
-    *,
-    policy_step: InferenceTierPolicyStep,
-    observation: InferenceObservation,
-    catalog: ActionCatalog | None,
-) -> bool:
-    """Return True when the ladder should stop after this completed step."""
-    if not policy_step.allow_ship_only_exact_early_stop:
-        return False
-    if catalog is None:
-        return False
-    best_solution = _best_merged_solution(state.merged_solutions)
-    if best_solution is None:
-        return False
-    if not _solution_qualifies_for_ship_only_exact_early_stop(
-        best_solution,
-        observation,
-        catalog,
-    ):
-        return False
-    state.ladder_complete = True
-    state.ladder_early_stop_reason = "ship_only_exact_early_stop"
-    return True
-
-
-def _maybe_no_new_exact_signatures_early_stop(
-    state: PolicyLadderState,
-    *,
-    added_combo_ids: frozenset[str],
-    added_aggregate_action_ids: frozenset[str],
-    new_exact_before_step: int,
-) -> bool:
-    """Return True when catalog growth was a noop and best exact is plausible enough."""
-    if len(state.merged_solutions) != new_exact_before_step:
-        return False
-    if not state.merged_solutions:
-        return False
-    if added_combo_ids or added_aggregate_action_ids:
-        return False
-    best_solution = _best_merged_solution(state.merged_solutions)
-    # Emptiness already gated above; _best_merged_solution only returns None for [].
-    assert best_solution is not None
-    thresholds = resolve_solver_thresholds()
-    if (
-        best_solution.objective_value
-        < thresholds.no_new_exact_signatures_early_stop_min_plausibility
-    ):
-        return False
-    state.ladder_complete = True
-    state.ladder_early_stop_reason = "no_new_exact_signatures"
-    return True
-
-
-class TierStepFinishMode(Enum):
-    """How ``_finish_tier_step`` closes the step after appending diagnostics."""
-
-    DIAGNOSTICS_ONLY = "diagnostics_only"
-    """Append diagnostics; caller owns ladder completion (cancel / invalid / abort)."""
-
-    SKIP = "skip"
-    """Advance index + ship-only early-stop (widen / raise skip paths)."""
-
-    COMPLETE = "complete"
-    """Advance index + ship-only and no-new-signatures early-stops (happy path)."""
-
-    BUDGET_STOP = "budget_stop"
-    """Advance index with no early-stop (tier allowance exhausted)."""
-
-
-def _finish_tier_step(
-    state: PolicyLadderState,
-    *,
-    policy_step: InferenceTierPolicyStep,
-    policy_step_index: int,
-    catalog: ActionCatalog | None,
-    turn: TurnInfo,
-    observation: InferenceObservation,
-    seed_count: int,
-    band_residual_2x: int | None,
-    step_started_at: float,
-    held_count_before: int,
-    newly_admitted: list[InferenceSolution],
-    collision_widen: CollisionHullWidenPlan | None = None,
-    prior_fleet_tech_raise: PriorFleetTechRaisePlan | None = None,
-    skipped: bool = False,
-    finish_mode: TierStepFinishMode = TierStepFinishMode.DIAGNOSTICS_ONLY,
-    tier_allowance_seconds: float | None = None,
-    reserved_for_later_seconds: float | None = None,
-    spendable_seconds: float | None = None,
-    added_combo_ids: frozenset[str] = frozenset(),
-    added_aggregate_action_ids: frozenset[str] = frozenset(),
-    new_exact_before_step: int | None = None,
-) -> None:
-    """Append tier diagnostics, then advance / early-stop per ``finish_mode``."""
-    state.step_diagnostics.append(
-        _policy_step_diagnostics(
-            policy_step=policy_step,
-            policy_step_index=policy_step_index,
-            catalog=catalog,
-            turn=turn,
-            observation=observation,
-            seed_count=seed_count,
-            band_residual_2x=band_residual_2x,
-            emission_fields=tier_emission_fields(
-                duration_ms=(time.monotonic() - step_started_at) * 1000.0,
-                held_count_before=held_count_before,
-                held_count_after=len(state.merged_solutions),
-                newly_admitted=newly_admitted,
-                time_limited=state.time_limited,
-                last_status=state.last_status,
-                skipped=skipped,
-                tier_allowance_seconds=tier_allowance_seconds,
-                reserved_for_later_seconds=reserved_for_later_seconds,
-                spendable_seconds=spendable_seconds,
-            ),
-            collision_widen=collision_widen,
-            prior_fleet_tech_raise=prior_fleet_tech_raise,
-        )
-    )
-    if finish_mode is TierStepFinishMode.DIAGNOSTICS_ONLY:
-        return
-
-    state.next_step_index = policy_step_index + 1
-
-    if finish_mode in (TierStepFinishMode.SKIP, TierStepFinishMode.COMPLETE):
-        if _maybe_early_stop_after_step(
-            state,
-            policy_step=policy_step,
-            observation=observation,
-            catalog=catalog,
-        ):
-            _annotate_last_step_early_stop(state)
-            return
-    if finish_mode is TierStepFinishMode.COMPLETE:
-        assert new_exact_before_step is not None
-        if _maybe_no_new_exact_signatures_early_stop(
-            state,
-            added_combo_ids=added_combo_ids,
-            added_aggregate_action_ids=added_aggregate_action_ids,
-            new_exact_before_step=new_exact_before_step,
-        ):
-            _annotate_last_step_early_stop(state)
-            return
-    # BUDGET_STOP: advance only -- no early-stop.
-
-    if state.next_step_index >= len(state.policy_steps):
-        state.ladder_complete = True
-
-
-def _make_incremental_admitter(
-    state: PolicyLadderState,
-    on_admitted: Callable[[InferenceSolution], None] | None,
-) -> Callable[[InferenceSolution], None]:
-    """Merge each solver solution into held top-K as soon as it is found."""
-
-    def admit(solution: InferenceSolution) -> None:
-        _merge_exact_solutions(
-            state.merged_solutions,
-            state.seen_signatures,
-            (solution,),
-            resolved_max_solutions=state.resolved_max_solutions,
-            on_admitted=on_admitted,
-        )
-
-    return admit
 
 
 def _abort_tier_step_on_seed_result(
@@ -552,7 +261,7 @@ def run_policy_ladder_tier_step(
         # Zero tier allowance (min=0 and nothing spendable), or cancel.
         if run.is_tier_only_stop():
             state.policy_steps_attempted.append(policy_step.id)
-            _finish_tier_step(
+            finish_tier_step(
                 state,
                 policy_step=policy_step,
                 policy_step_index=step_index,
@@ -606,7 +315,7 @@ def run_policy_ladder_tier_step(
             twins_fell_back=state.hull_collision_twins_fell_back,
         )
         if collision_widen.skipped:
-            _finish_tier_step(
+            finish_tier_step(
                 state,
                 policy_step=policy_step,
                 policy_step_index=step_index,
@@ -633,7 +342,7 @@ def run_policy_ladder_tier_step(
     )
     if prior_fleet_tech_raise is not None:
         if prior_fleet_tech_raise.skipped:
-            _finish_tier_step(
+            finish_tier_step(
                 state,
                 policy_step=prior_fleet_tech_raise.policy_step,
                 policy_step_index=step_index,
@@ -709,7 +418,7 @@ def run_policy_ladder_tier_step(
         added_aggregate_action_ids: frozenset[str] = frozenset(),
         new_exact_before_step: int | None = None,
     ) -> None:
-        _finish_tier_step(
+        finish_tier_step(
             state,
             policy_step=policy_step,
             policy_step_index=step_index,
